Backend/requirements.txt
flask
flask-cors
flask-jwt-extended
flask-bcrypt
pymongo
google-generativeai
python-docx
python-pptx